ipareq is now int 8

+++ src/Modules/InputData.F90
@@ -59,7 +59,7 @@ module inputData
     !           3-'infinite dilution parameters table (UNIFAC)'
     !           4-'GC-EOS parameters'
     !
-    integer(kind = int32) :: & 
+    integer(kind = int8) :: & 
     ! Some compilers struggle with these varibles (specially "ipareq") if they
     ! are int8
         & icalc = 0,&
